test-requirements.txt
pytest==3.7.0
pytest-mock
pytest-cov
pycodestyle>=2.3.1
astroid>=1.4.8
pylint>=1.7.1
tox>=2.3.1
mock
codecov
requests_kerberos
